+++ app/src/test/java/com/tasirin/vaultwardenhost/TlsCertTest.java
@@ -136,6 +136,8 @@ public class TlsCertTest {
         assertFalse(TlsCert.masaBelumTiba(hilang));
         File sampah = File.createTempFile("sampah", ".pem");
         Files.write(sampah.toPath(), "bukan-sertifikat".getBytes(StandardCharsets.US_ASCII));
+        assertEquals(-1, TlsCert.sisaMs(hilang));
+        assertEquals(-1, TlsCert.sisaMs(sampah));
         assertEquals(-1, TlsCert.daysLeft(sampah));
         assertFalse(TlsCert.masaBelumTiba(sampah));
         sampah.delete();